fl-models/mnist-federated-learning/requirements.txt
flwr~=1.4.0
openml~=0.13.1
concrete-ml